hwstructs: add errno to exceptions, raise ValueError on dyn size mismatch
We use IOError() for cases where `stream.read()` size is shorter than
expected, as to suggest that this may be a dynamic problem with that
stream. Yet `IOError()` expects an errno, pick 5=EIO there.
For dynamic vs static size, which is within the bounds of read buffer,
this is a computational error on the received data, use `ValueError` .